Introduce git-blamediff script
When working with git it is often a necessity to know which change set a
given line was changed in in order to merge something in or amend it in
some other way. In general, this task can be accomplished by git's
'blame' or 'annotate' commands.
It gets more interesting when one wants to annotate exactly what is in a
patch. That is, one wants to retrieve such an annotation for exactly
those lines present in a diff in order to generate fix up commits, for
instance.
This change introduces the git-blamediff script which can be used to
simplify this task. The script reads from standard input (where it
expects a diff), extracts the different hunks, and invokes git blame for
each of them. A sample invocation could look like this:
  $ git diff | git blamediff

In its current state, the annotation is always performed against HEAD
which makes sense in general for invocations of git 'diff' but not for
other commands such as 'show'.

Note: In order for git's 'blame' command to be able to directly
      understand the paths supplied you have to produce the diffs
      without a prefix. That is, the --no-prefix option should be passed
      to git's 'diff' command.
      Furthermore, in order for the script to work properly from
      directories other than the root of the git repository you will
      need to produce diffs with relative paths, i.e., pass in the
      --relative option as well.