Properly parse diffs containing continuation characters
When a file contains a trailing newline and this symbol is removed or
when it does not contain a trailing newline and such a symbol is added,
git emits a special line "\ No newline at end of file" that indicates
this change. As it stands, this line could not be parsed correctly for
it does not start with '+', '-' (both indicating an actual addition or
removal), or a space (indicating context).
This change adds support for the '\' symbol to the parser, causing it to
effectively ignore it.