Correctly handle diff changing a affecting a single line
If a patch adds or removes a file with a single line the count fields of
the diff header are left out because a count of one is the default. Our
diff parser could not handle this case and would report an unexpected
line.
This change adjusts the regular expression matching the diff header to
treat the count field as optional. If this field is missing, the assumed
count is one.