Properly flush stdout
When we print the file header to stdout Python might have an internal
buffer and so the output might not be visible immediately. That is fine
in principle but if we fork off another process which in turn prints
something to stdout and dies, this latter output is guaranteed to be
emitted to the user. If our Python instance is slow, the two prints
could be visible in the wrong order.
To fix this problem, this change flushes the stdout before forking off a
process to make sure the data is visible to the user in the intended
order.