Add end-to-end test
Although we test the diff parsing parts, there are some more bits to
git-blamediff that are not tested in an automated fashion.
To that end, this change adds an end-to-end test creating a git
repository, committing files, modifying them, and invoking git-blamediff
on the resulting diff. The output is then compared against the expected
one.