Correctly blame no longer existing files
If a file is removed it is no longer existant at the time of the git
blame invocation. Because the file no longer exists, the invocation will
fail.
This change adjust the invocation by inserting a double dash between the
standard parameters and the file name. By doing so, git is advised to
also check the file in the history of the repository.