Handle empty lines correctly
When a file contains an effectively empty line, a diff generated
incorporating this line as context does not have any visible character
in this particular line (not even a space as one would intuitively
expect).
This change adjust the diff parser to correctly handle this case. We
simply ignore such lines as they cannot induce a state change currently.